[Doc V3] Fixed header on scroll WIP
